añado user_detail, user, rol desde dto, respository y mapper

+++ src/main/java/com/diplomado/users/repositories/UserDetailRepository.java
@@ -3,5 +3,5 @@ package com.diplomado.users.repositories;
 import com.diplomado.users.domain.entities.UserDetail;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserDetailRepository extends JpaRepository<UserDetail,Integer> {
+public interface UserDetailRepository extends JpaRepository<UserDetail,Long> {
 }
